backend/requirements.txt
# Django
pytz==2018.7
Django==3.0.7
django-solo==1.1.3
django-filter==2.0.0
dj-static==0.0.6
django-hijack==2.1.4
django-rosetta==0.9.0
raven==6.3.0
django-cors-headers==2.2.0
django-import-export==1.1.0


# Rest framework
djangorestframework==3.8.0
djangorestframework-jwt==1.11.0

# Azure Storage
git+https://github.com/ragonzal/django-azure-backup.git@master#egg=django_azure_backup

# Celery
celery==4.3.0

# Linters and tests
pytest==4.0.1
pytest-django==3.4.4
coverage==4.5.2
pylint==2.2.2
pylint-django==2.0.4
Pillow==5.2.0
factory-boy==2.11.1
